requirements.txt
pypdf==6.9.1
pytest==9.0.2
litellm==1.82.4
jsonschema==4.26.0
jsonargparse==4.47.0
questionary==2.1.1
pdf2image==1.17.0
img2pdf==0.6.3
Pillow==11.1.0
PySide6>=6.6.0
asciidoc3==3.2.3
html2text==2025.4.15
python-Levenshtein==0.27.3
